refactor(main): 提取 fs-retry 共享模块 + 全仓 rmSync/renameSync 加固
- 新增 fs-retry.ts：导出 rmSyncWithRetry / renameWithRetry / RETRYABLE_FS_CODES
- agent-session-manager.ts：移除内联定义，改为从 fs-retry 导入
- agent-workspace-manager.ts：6 rmSync + 4 renameSync → retry 版本
- storage-service.ts：3 rmSync → rmSyncWithRetry
- migration-service.ts：6 rmSync → rmSyncWithRetry
